Move playAmbience from Game to Scene class
